Move biz(conference server) to /apps.
