--- configs/certs/cert.pem
@@ -1,26 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIEejCCAuKgAwIBAgIRAI714rb4Zaj9JZSPRyS9GYwwDQYJKoZIhvcNAQELBQAw
-gaUxHjAcBgNVBAoTFW1rY2VydCBkZXZlbG9wbWVudCBDQTE9MDsGA1UECww0ZHVh
-bndlaXdlaUBkdWFud2Vpd2VpZGVNYWNCb29rLVByby5sb2NhbCAo5q6157u05Lyf
-KTFEMEIGA1UEAww7bWtjZXJ0IGR1YW53ZWl3ZWlAZHVhbndlaXdlaWRlTWFjQm9v
-ay1Qcm8ubG9jYWwgKOautee7tOS8nykwHhcNMTkwNjAxMDAwMDAwWhcNMzAxMjIw
-MDQxODAxWjBXMScwJQYDVQQKEx5ta2NlcnQgZGV2ZWxvcG1lbnQgY2VydGlmaWNh
-dGUxLDAqBgNVBAsMI2R1YW53ZWl3ZWlAMTkyLjE2OC4xLjE1ICjmrrXnu7TkvJ8p
-MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEA3GvWVPRpW+IwskrJvmOB
-zjjIvi+6RVsVrcNWzum0vxJSfjUqYb7UmypD+C0BHZ7nyEG7Pbw3nMTASFum4/Pz
-HI++Chr/YJMqe7AUEsTeqwyJbX4wLse6WVEeD053Mi8gbmCYdWwCvdLArRkilYFS
-zZE/i1rUh1Qjj9sQR6jqlJKRqkOFf2EA2P+/TV5e8PqtqK9QC8nv/6iTUru+VxYU
-aEev+kiKMnSmFZi6SF3rPI+rv4DanwWHrAMIm5Pjozan4KvvxEkRP4670HWDdxzs
-pqL6Yy6rhir9dyjOcLMtLIrbmBjmxac6VUPSkiEUZXF765QhJwiJKj5IOUBESgQ4
-lwIDAQABo3IwcDAOBgNVHQ8BAf8EBAMCBaAwEwYDVR0lBAwwCgYIKwYBBQUHAwEw
-DAYDVR0TAQH/BAIwADAfBgNVHSMEGDAWgBTyCiopU0AMGvhZ2LHW9rv5UjFjQzAa
-BgNVHREEEzARgglsb2NhbGhvc3SHBH8AAAEwDQYJKoZIhvcNAQELBQADggGBAEP/
-hDz/GTKQZx3gzPfBC3Zl/mSJtTa1F0DniaKIgjBB/V4e/wom5avURBBdhnlDLhQm
-oqB+5LeT95QQnNO7EoyyxAcaf4fM91QH+TkIq1JGyIBF8Xc4h4xo/2KH93dUPa/m
-9M78IvO4mTU9T9q3/d/2EFHCrWckoUJokdko9ftnFUaWQ8nnD9tKAfXTZMwDP3R7
-frLELIRLojCF3d5ykgnNapoccWTjop6h9gbwZ+AHPOF/0Fs59SMJ9Uqx3OXYyNfq
-YJotreDsdS0fFdT4Vp64Tzh5m1Qo77BtFWCwiG9YQQNf00unOk7bViJyoubH96D6
-uQ/YBUrdgLuWJ7kZ9VxJ/aimmch6+VQ8j4xoayPlmhSFx9k2cE5h4ATnrFQze0y4
-6I5g9cpwNYnIB0IIidM0EyXaq/Vx1roWeSWVrG+V0FTEWxNLdAYchoxgxfBBQyKh
-qn9UiRiesrhUf/leHXSRFQT8wmznfN3I67/GANoJxLI7BJ3RlElGO2thssw5JA==
------END CERTIFICATE-----
