--- configs/certs/key.pem
@@ -1,28 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQDca9ZU9Glb4jCy
-Ssm+Y4HOOMi+L7pFWxWtw1bO6bS/ElJ+NSphvtSbKkP4LQEdnufIQbs9vDecxMBI
-W6bj8/Mcj74KGv9gkyp7sBQSxN6rDIltfjAux7pZUR4PTncyLyBuYJh1bAK90sCt
-GSKVgVLNkT+LWtSHVCOP2xBHqOqUkpGqQ4V/YQDY/79NXl7w+q2or1ALye//qJNS
-u75XFhRoR6/6SIoydKYVmLpIXes8j6u/gNqfBYesAwibk+OjNqfgq+/ESRE/jrvQ
-dYN3HOymovpjLquGKv13KM5wsy0situYGObFpzpVQ9KSIRRlcXvrlCEnCIkqPkg5
-QERKBDiXAgMBAAECggEBAIuguB4K0a8F4skefRvg1vp2mBt6Cmoi3+G56wSZoQik
-p6MRNK7EftNn+7R7dch4v+ToVp8t2qCxNgjgHiEqiKdgGwc1wO5Ba1j82kplcZQz
-Fd9JX2ZS2SX5L6yMjQ86MTYadJl1eWsb1TgRwwmn6XcU+obhW936G6jmITLzut2v
-qbvdLllCTGJzUjnRSgN+LBfu92N3tChYysUGeuKH/n2HG9zNmjq3mZc0Fs7fBXTz
-BkTUEzpuUifCQgZpHyjNI54SB4sTGheKxjpZsveatZ58sGjbKg/I397LRFg8H+ww
-OhKoPbMc5Aq4BqXL+I5BMRf9HOaHH6CRwd7d3tT43TECgYEA4UmRQddFC4ehguWa
-X3mXmODBpdPtHRDY+BirvzlCbMDVkp9HIClzX3VY+O7IBtzXGWpBqF4kXVNpGm4M
-sRMos7bzaNw7SW4nJlhNlk9aF1pXMw+wkHFB5Awyu2Iy1aoHYnmH6CkZF82K7E1Y
-55wBvwnB7GVR9I8me7zMVq4kk/kCgYEA+nhxoFRYrOBArQ8s3yfQZwjXmXFbo6Re
-tEUWKEowXgYEUAHZIoCfbMEbGH6RFh5qVJVG8OX4bYRGkt3+DV+UBzjfur2vtdt5
-qlE9bticTcoHJyMTiBepQpyZq11hLotP6fKURd5qVnQ8YmTnwd8jQ/ziZlyeztGo
-UBlSmbJlNQ8CgYEAu/uj/jl/PnCtH0kw9+vAyGr6isF8gB1XviTv4n8H4zFAxLit
-I7QStht34zdzpu5NecKFCNWNKGZ6M/iSFIt1EghdsSjbO5lzh3Z1Py/1b7kNO+ki
-nlQy8/QNvPfijRD0WoIoGlt4ubjj6+tafOBbQKoGq5ebvCJarTNLmiRzFxkCgYAJ
-jAG4vL5MiJeugP+S2e+lnb6qCJoXmLJLaL65dmp1iwErE4b1O7XX7VxdCKX7E9m1
-s6Lql7lqYzpVasS+OcBorhBJC+XgpqbyiPoweW0rweGX6kxaSZyX/2QUFhEYNxMq
-taZVK28eMHwEvBfhWeH8O21dvXYIKBGkS/LMDPKynQKBgQDN2goXeKIGqVGSBjeI
-vMsvc7Pv35zBv9T7pa2l+nrT0yQq4Bk0ZSnOxrGookuAX6tWxJ6HLXRPiT1An3a3
-C9xbBg3Qg+tzLy2mSBDtPQzbZd/+bwi9/zI5i/5FBOzWRHqdGbj+qNMUWUnsIjSo
-VflFEiCr9+YFjSkyHeTaO1EKRQ==
------END PRIVATE KEY-----
