+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/cloudwebrtc/nats-grpc v0.1.6
 	github.com/dgrijalva/jwt-go v3.2.0+incompatible
 	github.com/go-redis/redis/v7 v7.4.0
+	github.com/golang/protobuf v1.4.3
 	github.com/improbable-eng/grpc-web v0.13.0
 	github.com/jhump/protoreflect v1.8.2
 	github.com/nats-io/nats-server/v2 v2.1.9
